+++ src/components/SSF/SSF.test.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import { render, fireEvent } from '@testing-library/react'
+import { render, fireEvent, act } from '@testing-library/react'
 import SmartFilter from "./SSF"
 import { ReactComponent as ArrowSolidRight } from "../../assets/svg/ArrowSolidRight.svg"
 import { screen } from "@testing-library/dom"
@@ -31,37 +31,37 @@ const addFreeTextChip = value => {
 
 describe('<SSF />', () => {
   describe('auto complete menu', () => {
-    it('should open auto complete menu when input focus', () => {
+    it('should open auto complete menu when input focus', async () => {
       const { getByRole } = render(<SmartFilter fields={ fields } />)
       const input = getByRole('textbox')
-      input.focus()
+      await act(async () => input.focus())
       const autoComplete = getByRole('menu')
       expect(autoComplete).toBeTruthy()
     })
 
-    it('should open menu when click filter icon', () => {
+    it('should open menu when click filter icon', async () => {
       const { getByRole } = render(<SmartFilter fields={ fields } />)
       const filterIcon = getByRole('button')
-      filterIcon.click()
+      await act(async () => filterIcon.click())
       const autoComplete = getByRole('menu')
       expect(autoComplete).toBeTruthy()
     })
 
-    it('should show menu items that match typing', () => {
+    it('should show menu items that match typing', async () => {
       const { getByRole, getAllByRole } = render(<SmartFilter fields={ fields } />)
       const input = getByRole('textbox')
-      input.focus()
+      await act(async () => input.focus())
       let menuItems = getAllByRole('menuitem')
       expect(menuItems).toHaveLength(3)
-      fireEvent.change(input, { target: { value: 'Menu Item Text' } })
+      await act(async () => fireEvent.change(input, { target: { value: 'Menu Item Text' } }))
       menuItems = getAllByRole('menuitem')
       expect(menuItems).toHaveLength(1)
     })
 
-    it('should add text from menu to input and validate following inputs', () => {
+    it('should add text from menu to input and validate following inputs', async () => {
       const { getByRole, getAllByRole } = render(<SmartFilter fields={ fields } />)
       const input = getByRole('textbox')
-      input.focus()
+      await act(async () => input.focus())
       let menuItems = getAllByRole('menuitem')
       expect(menuItems).toHaveLength(3)
       fireEvent.click(menuItems[2])
@@ -76,8 +76,8 @@ describe('<SSF />', () => {
   })
 
   describe('events', () => {
-    it('onChange should return an array with chip objects containing chip field and value', () => {
-      const mockRegularChip  = 'Normal Chip'
+    it('onChange should return an array with chip objects containing chip field and value', async () => {
+      const mockRegularChip = 'Normal Chip'
       const mockFieldChip = 'Example Field'
       const onChange = chipsArr => {
         chipsArr.forEach((chip, index) => {
@@ -91,12 +91,14 @@ describe('<SSF />', () => {
           }
         })
       }
-      const { getAllByRole, getByRole } = render(<SmartFilter fields={ fields } onChange={ onChange } />)
-      addFreeTextChip(mockRegularChip)
-      const input = getByRole('textbox')
-      let menuItems = getAllByRole('menuitem')
-      fireEvent.click(menuItems[2])
-      addFreeTextChip(`${input.value}${mockFieldChip}`)
+      await act(async () => {
+        const { getAllByRole, getByRole } = render(<SmartFilter fields={ fields } onChange={ await onChange } />)
+        addFreeTextChip(mockRegularChip)
+        const input = getByRole('textbox')
+        let menuItems = getAllByRole('menuitem')
+        fireEvent.click(menuItems[2])
+        addFreeTextChip(`${input.value}${mockFieldChip}`)
+      })
     })
   })
 })