+++ src/hooks/UsePopoverPositioning/index.js
@@ -1,1 +1,0 @@
-export { default as usePopoverPositioning } from './usePopoverPositioning'