+++ src/hooks/index.js
@@ -1,4 +1,3 @@
 export { useClickOutsideListener } from './UseClickOutsideListener'
 export { useWindowDimensions } from './UseWindowDimensions'
-export { usePopoverPositioning } from './UsePopoverPositioning'
 export { usePrevious } from './UsePrevious'