+++ CHANGELOG.md
@@ -1,11 +1,13 @@
 ## Unreleased
 ### Fixed
 - TextField empty state bug
+- Positioning of the menu dropdown by using popper JS
 
 ## 1.0.0 - 06-09-2020
 
 ### Breaking Changes :boom:
 - Added label prop to textField
+- Removed AttachedAxis and changed preferOpenDirection to match the directions PopperJS uses
 
 ### Changed
 - Changed textfield to be uncontrolled
@@ -19,6 +21,7 @@
 - New Table.
 - New Text Field.
 - New Wizard Dialog.
+- Added default behavior that doesn't require isOpen in Menu
 
 <!-- 
 Sections: